+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file. This project adheres to
 [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Fixed
+* importing module issue
+
+### Added
+* Support for custom portal endpoint 
+
 ## [2.0.0] - 2023-03-30
 ### Added
 * Support for `bbox` query parameter for Arcgis portal search